requirements/base.txt
Django==1.11.10
django-cors-headers==2.2.0
django-environ==0.4.4
django-filter==1.1.0
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
Pillow==5.0.0
PyJWT==1.6.1
pytz==2018.3
six==1.11.0
